microservices/auth-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
sqlalchemy==2.0.23
sqlmodel==0.0.16
emails==0.6
jinja2==3.1.4
redis==5.0.1
python-dotenv==1.0.0
cryptography==41.0.7
bcrypt==4.0.1
pyjwt==2.8.0
typing-extensions>=4.8.0
anyio<4.0.0,>=3.7.1
pydantic-core==2.10.1
cffi>=1.12